pyproject.toml
[project]
name = "ghostwriter"
version = "0.1.0"
description = "Unified messaging application manager with intelligent processing"
authors = [{ name = "Nick L.", email = "lesseosnick@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "cryptography>=41.0.0",
]

[tool.poetry]
packages = [{ include = "ghostwriter", from = "src" }]
include = ["src/ghostwriter/py.typed"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
ruff = "^0.12.10"
mypy = "^1.0.0"
pre-commit = "^3.0.0"
types-requests = "^2.32.4.20250809"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ghostwriter = "ghostwriter.main:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "PIE", # flake8-pie
    "SIM", # flake8-simplify
]
ignore = ["E203"]  # whitespace before ':'

[tool.ruff.lint.isort]
known-first-party = ["ghostwriter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers --disable-warnings"
asyncio_mode = "auto"
markers = [
    "asyncio: mark test as async",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "."
exclude = ["build", "dist", "__pycache__"]
ignore_missing_imports = false
